feat: add map infos

+++ src/index.ts
@@ -7,6 +7,7 @@ import { PathLike } from 'fs';
 import { updateEnemies } from './updateEnemies.mjs';
 import { updateTroops } from './updateTroops.mjs';
 import { updateStates } from './updateStates.mjs';
+import { updateMapInfos } from './updateMapInfos.mjs';
 
 const program = new Command();
 
@@ -23,11 +24,14 @@ const oldPath = options.oldPath ? path.resolve(options.oldPath) : 'Old';
 const newPath = options.newPath ? path.resolve(options.newPath) : 'New';
 
 const oldDatabasePath = path.join(oldPath, 'RPG_RT.edb');
+const oldMapTreePath = path.join(oldPath, 'RPG_RT.emt');
 const actorsPath = path.join(newPath, 'Actors.json');
 const skillsPath = path.join(newPath, 'Skills.json');
 const enemiesPath = path.join(newPath, 'Enemies.json');
 const troopsPath = path.join(newPath, 'Troops.json');
 const statesPath = path.join(newPath, 'States.json');
+const mapInfosPath = path.join(newPath, 'MapInfos.json');
+
 
 /**
  * 
@@ -45,9 +49,10 @@ async function exists(f: PathLike) {
 
 async function main() {
   try {
-    if (await exists(oldDatabasePath)) {
+    if (await exists(oldDatabasePath) && await exists(oldMapTreePath)) {
       console.log('Updating actors');
-      const oldDatabaseXml = await fs.readFile(oldDatabasePath,"utf-8");
+      const oldDatabaseXml = await fs.readFile(oldDatabasePath, "utf-8");
+      const oldMapTreeXML = await fs.readFile(oldMapTreePath, 'utf-8');
       const updatedActorsJson = await updateActors(oldDatabaseXml);
       await fs.writeFile(actorsPath, updatedActorsJson, { encoding: 'utf-8' });
       console.log('Updating skills');
@@ -68,14 +73,17 @@ async function main() {
       await fs.writeFile(statesPath, updatedStatesJson, {
         encoding: 'utf-8'
       });
+      console.log('Updating map infos');
+      const updatedMapInfoJson = await updateMapInfos(oldMapTreeXML);
+      await fs.writeFile(mapInfosPath, updatedMapInfoJson, {
+        encoding: 'utf-8'
+      });
     } else {
       console.error(
-        'Error: RPG_RT.edb not found. Please check the provided file path and spelling.'
+        'Error: RPG_RT.edb or RPG_RT.emt not found. Please check the provided file path and spelling.'
       );
       process.exit(-1);
     }
-
-    // other update calls
   } catch (ex) {
     console.error('An unexpected error occurred.');
     console.log(ex);
@@ -83,5 +91,4 @@ async function main() {
 }
 
 main();
-
